[rename] : memberInfo DTO 위치 이동
