+++ src/main/java/com/dnd/gongmuin/chat_inquiry/dto/ChatInquiryResponse.java
@@ -2,7 +2,7 @@ package com.dnd.gongmuin.chat_inquiry.dto;
 
 import com.dnd.gongmuin.chat_inquiry.domain.InquiryStatus;
 import com.dnd.gongmuin.member.domain.JobGroup;
-import com.dnd.gongmuin.question_post.dto.response.MemberInfo;
+import com.dnd.gongmuin.member.dto.response.MemberInfo;
 import com.querydsl.core.annotations.QueryProjection;
 
 public record ChatInquiryResponse(
